+++ download.sh
@@ -5,8 +5,12 @@ escape_char=$(printf '\u1b')
 tmp='/var/tmp'
 jq="${tmp}/jq"
 repo='https://api.github.com/repos/m3l6h/utilities'
-# repo='https://api.github.com/repos/stedolan/jq'
 releases="${repo}/releases"
+height="$(tput lines)"
+width="$(tput cols)"
+
+instructions=("Select the utilities you would like to install" "Navigate with the arrow keys" "Press A to select the latest version of each utility" "Press S to toggle a particular utility" "Press D to select all utilities" "Press Q to quit this menu" "Press Enter to confirm your selection" " ")
+offset="${#instructions[@]}"
 
 function gcurl {
   [ -z "$authentication" ] && \
@@ -25,29 +29,77 @@ function get_util {
   echo "${tmp}/${2}.tar.gz"
 }
 
-function print_scr {
-  clear
-  printf "Select the utilities you would like to install\n"
-  printf "Navigate with the arrow keys\n"
-  printf "Use A to select all utilities\n"
-  printf "Use S to toggle a particular utility\n"
-  printf "Use D to select all utilities\n"
-  printf "Use Q to quit this menu\n"
-  printf "Use Enter to confirm your selection\n"
-  offset=7
+# Expects the following:
+#   $1 - String to print
+#   $2 - (Optional) Color
+function println {
+  local strlen="${#1}"
+  local offset=1 numlines=0 substr cutpoint
 
-  numlines=0
+  [ -n "$2" ] && echo -en "$2"
 
-  local util
-  for util in "${names[@]}"; do
-    printf "[ ] ${util}"
-    "${prerelease[$numlines]}" && printf ' (prerelease)'
-    printf '\n'
+  while true; do
+    cutpoint=$((offset + width - 1))
+
+    while [ "$cutpoint" -ge "$offset" ]; do
+      [ "$cutpoint" -ge "$strlen" ] && break
+      [ "$(cut -c "${cutpoint}-${cutpoint}" <<<"$1")" = ' ' ] && break
+      cutpoint=$((cutpoint - 1))
+    done
+
+    [ "$cutpoint" -lt "$offset" ] && cutpoint=$((offset + width - 1))
+
+    substr="$(cut -c "${offset}-${cutpoint}" <<<"$1")"
+
+    [ -z "$substr" ] && break
+    printf "${substr}"
     numlines=$((numlines + 1))
+    offset=$((offset + cutpoint))
+    [ "$offset" -le "$strlen" ] && echo
   done
 
-  line=1
-  echo -en "\033[$((offset + line));2H"
+  [ -n "$2" ] && echo -en "\e[0m"
+
+  return "$numlines"
+}
+
+# Expects the following:
+#   $1 - The line the user is on
+function print_list {
+  if [ "$height" -le "$offset" ]; then
+    clear
+    println "Terminal window is not tall enough!\n"
+    exit 1
+  fi
+
+  local line="$1"
+  local space=$((height - offset))
+  local hidden=$((line - space))
+  [ "$hidden" -lt 0 ] && hidden=0
+  local i="$hidden" util count=1 str
+
+  while [ "$i" -lt "$numlines" ]; do
+    util="${names[i]}"
+    str="[$($(${toggles[i]}) && echo 'x' || echo ' ')] ${util}$("${latest[$i]}" && echo ' (latest)')$("${prerelease[$i]}" && echo ' (prerelease)')"
+    [ "${#str}" -gt "$width" ] && str="$(cut -c "1-$((width-3))" <<<"$str")..."
+    println "$str"
+    i=$((i + 1))
+    count=$((count + 1))
+    [ "$count" -gt "$space" ] && break || echo
+  done
+
+  echo -en "\033[$((offset + line - hidden));2H"
+}
+
+function print_scr {
+  clear
+  local instruction
+  offset=0
+  for instruction in "${instructions[@]}"; do
+    println "${instruction}\n"
+    local lines="$?"
+    offset=$((offset + lines))
+  done
 }
 
 function exit_handler {
@@ -57,26 +109,43 @@ function exit_handler {
 }
 
 function main {
+  if [ "$width" -lt 30 ]; then
+    println "Terminal must be at least 30 characters wide!\n"
+    exit 1
+  fi
+
   get_jq
   local releases="$(gcurl "$releases")"
   IFS=$'\n' names=($("$jq" -r '.[].name' <<<"$releases"))
   IFS=$'\n' prerelease=($("$jq" -r '.[].prerelease' <<<"$releases"))
   IFS=$'\n' urls=($("$jq" -r '.[].assets | .[0].browser_download_url' <<<"$releases"))
-  local toggles=( )
+  numlines="${#names[@]}"
+  toggles=( )
+  latest=( )
 
-  for dump in "${names[@]}"; do toggles+=( false ); done
+  local found_latest name
+  for name in "${names[@]}"; do
+    toggles+=( false )
+    name="$(awk '{ print $1; }' <<<"$name")"
+    if grep -q "$name" <<<"$found_latest"; then
+      latest+=( false )
+    else
+      latest+=( true )
+      found_latest="${found_latest}:${name}"
+    fi
+  done
 
-  local key
+  local key line=1
   print_scr
+  print_list "$line"
   while read -sn1 key; do
     [ "$key" = "$escape_char" ] && read -sn2 key
     case "$key" in
       'a')
         local i=0
         for t in "${toggles[@]}"; do
-          toggles[$i]=true
+          "${latest[i]}" && toggles[$i]=true
           i=$((i+1))
-          echo -en "\033[$((i + offset));2Hx"
         done
       ;;
       'd')
@@ -84,20 +153,19 @@ function main {
         for t in "${toggles[@]}"; do
           toggles[$i]=false
           i=$((i+1))
-          echo -en "\033[$((i + offset));2H "
         done
       ;;
       'q') exit_handler 'Download aborted\n' ;;
       's')
         $(${toggles[$((line - 1))]}) && toggles[$((line - 1))]=false || toggles[$((line - 1))]=true
-        $(${toggles[$((line - 1))]}) && printf 'x' || printf ' '
       ;;
       '[A') line=$((line - 1)) && [ 1 -ge "$line" ] && line=1 ;;
       '[B') line=$((line + 1)) && [ "$numlines" -lt "$line" ] && line="$numlines" ;;
       '') break ;;
       *) ;;
     esac
-    echo -en "\033[$((line + offset));2H"
+    echo -en "\033[$((offset + 1));1H"
+    print_list "$line"
   done
 
   clear