+++ upgrade.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+echo "Upgrading to $(<"./data/version")..."